refactor: separate components in different files
