+++ src/pages/Dashboard/Hotel/index.js
@@ -2,24 +2,8 @@ import { useContext, useState, useEffect } from 'react';
 import styled from 'styled-components';
 import UserContext from '../../../contexts/UserContext';
 import { getHotels } from '../../../services/hotelApi';
+import Container from './Container';
 import Typography from '@material-ui/core/Typography';
-
-function HotelComp({ value, selected, setSelected }) {
-  function handleClick() {
-    if(selected.id !== value.id) {
-      setSelected({ ...value });
-    }
-  }
-
-  return (
-    <UniqueHotel hotelid={value.id} selectedid={selected.id} onClick={handleClick}>
-      <div>
-        <img src={value.image} alt='Hotel' />
-        <span>{value.name}</span>
-      </div>
-    </UniqueHotel>
-  );
-}
 
 export default function Hotel() {
   const [hotels, setHotels] = useState([]);
@@ -28,7 +12,7 @@ export default function Hotel() {
 
   useEffect(() => {
     getHotels(userData.token).then((res) => {
-      setHotels([...hotels, res.data]);
+      setHotels([...res.data]);
     });
   }, []);
 
@@ -52,7 +36,7 @@ export default function Hotel() {
           </ChooseHotel>
           <Hotels>
             {hotels.map((value, index) => (
-              <HotelComp value={value} selected={selected} setSelected={setSelected} key={index} />
+              <Container value={value} selected={selected} setSelected={setSelected} key={index} />
             ))}
           </Hotels>
         </>
@@ -91,36 +75,4 @@ const Hotels = styled.div`
   column-gap: 19px;
   row-gap: 19px;
   width: 100%;
-`;
-
-const UniqueHotel = styled(Typography)`
-  background-color: ${({ hotelid, selectedid }) => hotelid === selectedid ? '#FFEED2' : '#EBEBEB'};
-  display: flex;
-  justify-content: center;
-  width: 196px;
-  height: 264px;
-  border-radius: 10px;
-  padding: 12px;
-  cursor: pointer;
-
-  & img {
-    object-fit: cover;
-    width: 168px;
-    height: 109px;
-    border-radius: 5px;
-  }
-  
-  & > div {
-    width: 100%;
-    row-gap: 8px;
-    font-size: 18px;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-  }
-  & > div > span {
-    display: flex;
-    width: 100%;
-    justify-content: left;
-  }
 `;
